long

+++ holberton.h
@@ -5,7 +5,7 @@
 #include <stdio.h>
 #include <unistd.h>
 int _putchar(char c);
-int print_number(int x); 
+int print_number(long x);
 int suich(const char *format, va_list list);
 int _printf(const char *format, ...);
 #endif
